WooPayments: address A4A MSD dashboard review feedback
- Use a single createInterpolateElement call for the add-to-site modal copy
  instead of splitting the sentence across multiple __() fragments.
- Drop the baked-in enabled param and redundant TanStack defaults from
  agencyWooPaymentsCommissionsQuery; callers pass enabled instead.
- Move the generic fetchSiteTestConnection / siteTestConnectionQuery out of
  agency-woopayments into the site-jetpack-connection resource.
- Format consolidated-stats currency and dates with a fixed en-US locale so
  they don't drift to the viewer's WordPress.com or browser locale (A4A is
  English-only for now).
