+++ packages/api-core/src/site-jetpack-connection/fetchers.ts
@@ -17,3 +17,11 @@ export async function fetchJetpackConnectionHealth(
 		apiNamespace: 'wpcom/v2',
 	} );
 }
+
+export async function fetchSiteTestConnection( blogId: number ): Promise< boolean > {
+	const response = await wpcom.req.get( {
+		apiNamespace: 'rest/v1.1',
+		path: `/jetpack-blogs/${ blogId }/test-connection`,
+	} );
+	return response?.connected !== false;
+}
